change return type to float

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -2,10 +2,10 @@
 -- Function SafeDiv prevents zerodivision error
 DELIMITER $$
 CREATE FUNCTION SafeDiv(a INT, b INT)
-RETURNS INT
+RETURNS FLOAT
 DETERMINISTIC
 BEGIN
-    DECLARE result INT;
+    DECLARE result FLOAT;
     IF b <= 0 THEN SET result = 0;
     ELSE SET result = a / b;
     END IF;
